backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
pydantic==2.6.3
python-jose==3.3.0
cassandra-driver==3.28.0